Refactoring in progress
Started to refactor .parse() method, by splitting it in multiple smaller methods

Changed the structure of PATTERNS so it's easier to reference any pattern by the name, before there were some magic numbers in parse function

Changed what is exposed from the package, only the class that the user can instantiate.

Added a refresh mechanism so the class can be used in a singleton manner.